Add support for custom cwd and initial impl for start
This pulls most of the Java running code out into functions so the logic
can be shared between start and run.

The start command is not yet working as the Paper server needs to be
modified as well to support this, namely it needs to be modified to
handle stdout, stderr, and stdin before anything tries to use them. As
part of the daemonization process, these files are closed, which causes
the JVM to crash.