Define extern functions for IPC message queues manually
This is an unfortunate requirement to allowing development on macOS.
Linux will always be the only target for this project, but I like
developing for it on macOS as well, which being Unix also supports these
functions. However, and I have no clue why, these functions are not
exported by libc on macOS, so we have to extern them ourselves. They
still work fine in practice though when manually added.