Implement JNI interfaces to move away from JNA
Using JNA is a much simpler implementation, but comes with the burden of
requiring JNA, and potentially conflicting with other libraries which
may require a different JNA version. Using JNI requires managing the
shared library somehow, which is accomplished in this commit by
implementing the JNI functions in a separate module.

The separate module, paperd-jni, is built before paperd and included as
a resource in paperd. paperd extracts it to disk before starting the
server in daemon mode.