Improve build system and stop handling
The building and embedding of paperd-jni is now done with a build.rs
file which is natively executed by Cargo. Cargo executes this file and
uses it's output to modify the build. I call Cargo again to build
paperd-jni and then output a command to Cargo to set the environment
variable to allow paperd to embed the compressed binary.

This has the downside of not allowing the log output of the paperd-jni
build be visible, but it has the benefit of significantly improving
build times as it is built in parallel with the other dependency builds.
It also has the benefit of being native with Cargo, which means
debugging and other tooling work out of the box now and no more hacky
workarounds in a shell script are needed.

The other change is with the Docker "publish" artifact. For
maintainability and more reproducable builds, the Dockerfile in this
repo is a more specific definition of what a release build is. I also
configured the Docker build to statically link ncurses which should
hopefully prevent people from having issues with ncurses not linking
properly when they download binary builds.

I also improved how messages are handled while paperd is stopping by
adding support for the `stop` AtomicBool in the console section to be
passed in. This allowed me to add stop checks in a lot more places and
lets me cleanup message queue resources cleaner than before. Also I
removed error messages from queue deletion, since it doesn't really
matter as we're done with the queue anyways.